requirements.txt
Click==7.0
dateparser==0.7.0
Flask==1.0.2
Flask-Assets==0.12
itsdangerous==0.24
Jinja2==2.10
jsmin==2.2.2
MarkupSafe==1.0
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
regex==2018.8.29
simplejson==3.16.0
six==1.11.0
tzlocal==1.5.1
webassets==0.12.1
Werkzeug==0.14.1
